requirements.txt
Django==4.2.14
django-model-utils==4.5.1
django-extensions==3.2.3
datasets==2.20.0
textstat==0.7.4
spacy==3.7.5
python-decouple==3.8
pandas==2.2.2
langdetect==1.0.9
gensim==4.3.3
scikit-learn==1.5.1
pymfe==0.4.3
matplotlib==3.9.1.post1
seaborn==0.13.2
torch==2.4.0
pymoo==0.6.1.3
pythae==0.1.2
notebook==7.2.1
